### 2019-08-14 #1647 @whedon
[PRE REVIEW]: deepCR: Cosmic Ray Rejection with Deep Learning
**Submitting author:** @kmzzhang (<a href="http://orcid.org/0000-0002-9870-5695">Keming Zhang</a>)
**Repository:** <a href="https://github.com/profjsb/deepCR" target ="_blank">https://github.com/profjsb/deepCR</a>
**Version:** 0.1.5
