### 2019-08-14 aceb356
added bib file for JOSS submission
