### 2019-08-14 #1647 @kmzzhang
Hello,

I have just included paper.bib in the repo.

Thanks,
Keming
