### 2019-08-14 #1647 @kmzzhang
@terrytangyuan, may I suggest nilesh-patil, DanielLenz, shreyasbapat as potential reviewers? They have listed both deep learning and astronomy as their expertise. Additionally, I don’t see the figure in the generated paper proof. Is the figure format wrong? Thanks.
