### 2019-08-14 #1647 @danielskatz
I wonder if the path to the image is a problem?  @arfon - can you help with this?  (since we can't see the generate pdf log to see any error msgs)
