### 2019-08-15 #1647 @DanielLenz
I'd be happy to review in general. However, the work has already been described and published in [Zhang & Bloom (2019, submitted to AAS)](https://arxiv.org/abs/1907.09500), and will undergo peer review there as well.
