### 2019-08-15 #1647 @arfon
> I wonder if the path to the image is a problem? @arfon - can you help with this? (since we can't see the generate pdf log to see any error msgs)

@kmzzhang - can you update the URL of the image to: `https://raw.githubusercontent.com/profjsb/deepCR/master/imgs/postage-sm.jpg` rather than using the relative path?

> I'd be happy to review in general. However, the work has already been described and published in Zhang & Bloom (2019, submitted to AAS), and will undergo peer review there as well.

Also, can you confirm @kmzzhang that this is a joint submission with AAS?
