### 2019-08-15 #1647 @arfon
> Yes, this is a joint submission with AAS.

👍 thanks for confirming.

/ cc @crawfordsm
