### 2019-08-15 #1651 @whedon
[REVIEW]: deepCR: Cosmic Ray Rejection with Deep Learning
**Submitting author:** @kmzzhang (<a href="http://orcid.org/0000-0002-9870-5695">Keming Zhang</a>)
**Repository:** <a href="https://github.com/profjsb/deepCR" target ="_blank">https://github.com/profjsb/deepCR</a>
**Version:** v0.2.0
