### 2019-08-16 profjsb/deepCR#8 @DanielLenz
Instructions to verify installation
Hi, I have the pleasure to review deepCR for your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/1651#event-2561054598).

I'll likely have a couple of smaller comments and questions during the next few days, and will raise separate issues to keep your github issues clean.

Could you add a quick paragraph in the README that allows the user to verify the installation, i.e. execute the tests locally? Ideally, you could just grab the matching lines from your travis config (`pip install pytest-XYZ; python setup.py test`).
