### 2019-08-16 profjsb/deepCR#9 @DanielLenz
Adding a CONTRIBUTING?
Hi, I have the pleasure to review deepCR for your [JOSS submission](https://github.com/openjournals/joss-reviews/issues/1651#event-2561054598).

I'll likely have a couple of smaller comments and questions during the next few days, and will raise separate issues to keep your github issues clean.

I really appreciate your paragraph on contributing in the README; do you  think it's worth considering to expand this to a full [CONTRIBUTING.md](https://help.github.com/en/articles/setting-guidelines-for-repository-contributors)? I think it would add value, is easier to find, and gives you the opportunity to provide more detailed guidelines and information on how to contribute.
