### 2019-08-16 #1651 @DanielLenz
I really enjoyed the package and the paper, thanks for providing that! I'll wrap up the review shortly.

In the meantime, could you add the software acknowledgements that you have included in our AAS submission here as well?
