+++ paper.bib
@@ -33,3 +33,119 @@ archivePrefix = "arXiv",
 booktitle = {NIPS-Workshop},
      year = {2017}
 }
+
+@ARTICLE{vdw11,
+author={S. {van der Walt} and S. C. {Colbert} and G. {Varoquaux}},
+journal={Computing in Science Engineering},
+title={The NumPy Array: A Structure for Efficient Numerical Computation},
+year={2011},
+volume={13},
+number={2},
+pages={22-30},
+keywords={data structures;high level languages;mathematics computing;numerical analysis;numerical computation;numpy array;numerical data;high level language;Python programming language;Arrays;Numerical analysis;Performance evaluation;Computational efficiency;Finite element methods;Vector quantization;Resource management;Python;NumPy;scientific programming;numerical computations;programming libraries},
+doi={10.1109/MCSE.2011.37},
+ISSN={1521-9615},
+month={March},}
+
+@inproceedings{astrodrizzle,
+   author = {{Hack}, W.~J. and {Dencheva}, N. and {Fruchter}, A.~S. and {Armstrong}, A. and
+	{Avila}, R. and {Baggett}, S. and {Bray}, E. and {Droettboom}, M. and
+	{Dulude}, M. and {Gonzaga}, S. and {Grogin}, N.~A. and {Kozhurina-Platais}, V. and
+	{Lucas}, R.~A. and {Mack}, J. and {MacKenty}, J. and {Petro}, L. and
+	{Pirzkal}, N. and {Rajan}, A. and {Smith}, L.~J. and {Sontag}, C. and
+	{Ubeda}, L.},
+    title = "{AstroDrizzle: More than a New MultiDrizzle}",
+booktitle = {American Astronomical Society Meeting Abstracts \#220},
+     year = 2012,
+   series = {American Astronomical Society Meeting Abstracts},
+   volume = 220,
+    month = may,
+      eid = {135.15},
+    pages = {135.15},
+   adsurl = {https://ui.adsabs.harvard.edu/abs/2012AAS...22013515H},
+  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@misc{scipy,
+  author =    {Eric Jones and Travis Oliphant and Pearu Peterson and others},
+  title =     {{SciPy}: Open source scientific tools for {Python}},
+  year =      {2001},
+  howpublished = {\url{http://www.scipy.org/}},
+  note = {Accessed: 2019-07-20}
+}
+
+@ARTICLE{matplotlib,
+   author = {{Hunter}, J.~D.},
+    title = "{Matplotlib: A 2D Graphics Environment}",
+  journal = {Computing in Science and Engineering},
+ keywords = {Python, Scripting languages, Application development, Scientific programming  },
+     year = 2007,
+    month = may,
+   volume = 9,
+    pages = {90-95},
+      doi = {10.1109/MCSE.2007.55},
+   adsurl = {https://ui.adsabs.harvard.edu/abs/2007CSE.....9...90H},
+  adsnote = {Provided by the SAO/NASA Astrophysics Data System}
+}
+
+@misc{astroscrappy,
+  author       = {Curtis McCully and
+                  Steve Crawford and
+                  Gabor Kovacs and
+                  Erik Tollerud and
+                  Edward Betts and
+                  Larry Bradley and
+                  Matt Craig and
+                  James Turner and
+                  Ole Streicher and
+                  Brigitta Sipocz and
+                  Thomas Robitaille and
+                  Christoph Deil},
+  title        = {astropy/astroscrappy: v1.0.5 Zenodo Release},
+  month        = nov,
+  year         = 2018,
+  doi          = {10.5281/zenodo.1482019},
+  url          = {https://doi.org/10.5281/zenodo.1482019}
+}
+
[…]
