### 2019-08-17 #1651 @shreyasbapat
Can the network architecture diagram be also included in the paper? Or maybe in the README as  it is in the https://arxiv.org/pdf/1907.09500.pdf ?
