### 2019-08-21 1c134ff
added network figure in JOSS paper
