### 2019-08-21 #1651 @kmzzhang
@shreyasbapat I've added the network architecture figure in the paper. I also added both figure caption to the main text because they're not showing up as figure captions. Please let me know if this is the optimal way to do it. Thanks!
