### 2019-08-21 #1651 @DanielLenz
Thanks for adding the figures! I'm sure there a more optimal way to adjust the alignment/spacing and the caption; do you have any pointers, @terrytangyuan ?
