### 2019-08-21 #1651 @terrytangyuan
cc @openjournals/joss-eics who might have a better idea on how to fix it.
