### 2019-08-21 #1651 @kyleniemeyer
@kmzzhang if you put the figures on their own lines, the positioning should be better; you can also add captions:
```markdown
``deepCR`` is a Python package ...
... at left) is concatenated with the original image as the input.

![Figure 1](https://raw.githubusercontent.com/profjsb/deepCR/master/imgs/network.png)

Since ``deepCR`` ...
...pixels replaced with deepCR predictions  (last row).

![Figure 2](https://raw.githubusercontent.com/profjsb/deepCR/master/imgs/postage-sm.jpg)

In the paper accompanying ``deepCR`` [@deepcr], the authors showed that ...
```
