### 2019-08-21 #1651 @shreyasbapat
Hi @kmzzhang ,

Thanks for including the figure, but can we change the caption to both the figures and make the alignment a little better?
