### 2019-08-21 #1651 @shreyasbapat
> I also added both figure caption to the main text because they're not showing up as figure captions

Oh I see :/
