### 2019-08-21 #1651 @DanielLenz
Excellent, thanks for fixing this so quickly.

Last quick comment: Could you separate the software acknowledgements with a delimiter (semicolon, comma)?

Other than that, I'm happy to say that I recommend this for publication!
