### 2019-08-30 #1651 @crawfordsm
For reference, the corresponding paper has been accepted at the AAS and the AAS DOI for the paper is 10.3847/1538-3881/ab3c53
