### 2019-08-30 #1651 @arfon
> Hi @terrytangyuan, could you please let us know the next steps? Thanks.

@kmzzhang - can you tell me what AAS journal you submitted the other paper to? There’s a small amount of metadata we need you to add to the `paper.md` file before accepting the JOSS paper (this links the two papers).
