### 2019-09-01 #1651 @kmzzhang
@arfon It is accepted to ApJ
