### 2019-09-01 b4f4e7d
Update paper.md
