### 2019-09-01 #1651 @arfon
@kmzzhang - can you merge this PR please? https://github.com/profjsb/deepCR/pull/10
