### 2019-10-28 #1651 @kmzzhang
@arfon I just realized that the title in the paper is incorrect. It should be “deepCR: Cosmic Ray Rejection with Deep Learning,” while the published paper has the word “Ray" missing. I may have omitted it in the paper source file — the title in the paper submission metadata is correct. Is there a way to fix this error?
